pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autoflow-srxn"
version = "0.1.0"
dependencies = [
    "ase",
    "pymatgen",
    "numpy",
    "scipy",
    "matplotlib",
    "pandas",
    "tqdm",
    "pyyaml",
    "phonopy",
    "spglib",
    "seekpath"
]

[project.optional-dependencies]
mace = [
    "torch",
    "torch-geometric",
    "mace-torch"
]
sevennet = [
    "sevenn"
]

